[LeetCode] 1438. Longest Continuous Subarray With Absolute Diff Less Than or Equal to Limit
We can have two pointers: left and right, and then find the longest subarray for every right pointer (iterate it) by shrinking left pointer. And return the longest one among them.

Everytime the right pointer reaches a new position, we need to dequeue the "tail" element who is smaller than the nums[right]. Since, those "old small tail" elements will never be the range maximum from now on. After "clean up" the "old small tail" elements, add nums[right] into the deque, and then, the head of deque is the current maximum.

Same for the min Deque. Move right poniter by 1, and clean up "old big tail" elements, add nums[right], the head of deque is the current minimum.

What we should do next is to shrink left pointer because of limit. If current.max - current.min > limit. We should move the left pointer. Accordingly, we need to update our min max deques. If head of max deque is equal to the nums[left], that means, it is the one, we need to dequeue it, since we are gonna move the left pointer by 1.