[LeetCode] 946. Validate Stack Sequences
Once top of stack is the same as pop, we need to pop it because if we don't do it, other values will be pushed to stack and the order will not be the same, ie, it can not pop at such order any more. Greedy pop according to the pop array and check if the stack is empty at the end.